Adding sync endpoint

+++ src/main/java/com/dfernandezyopla/PianoReminder/Game/Services/QuestionServiceImpl.java
@@ -1,19 +1,26 @@
 package com.dfernandezyopla.PianoReminder.Game.Services;
 
 import com.dfernandezyopla.PianoReminder.Exceptions.ChordNotFoundException;
+import com.dfernandezyopla.PianoReminder.Exceptions.EntityConstrainedException;
 import com.dfernandezyopla.PianoReminder.Exceptions.HistoryQuestionNotFoundException;
 import com.dfernandezyopla.PianoReminder.Exceptions.NoteNotFoundException;
 import com.dfernandezyopla.PianoReminder.Game.DTOs.ChordRequestDTO;
-import com.dfernandezyopla.PianoReminder.Game.Entities.Chord;
-import com.dfernandezyopla.PianoReminder.Game.Entities.HistoryQuestion;
-import com.dfernandezyopla.PianoReminder.Game.Entities.Note;
+import com.dfernandezyopla.PianoReminder.Game.Entities.*;
+import com.dfernandezyopla.PianoReminder.Game.Entities.Sync.EntityToSync;
+import com.dfernandezyopla.PianoReminder.Game.Entities.Sync.EntityType;
+import com.dfernandezyopla.PianoReminder.Game.Entities.Sync.SyncType;
 import com.dfernandezyopla.PianoReminder.Game.Repositories.ChordRepository;
+import com.dfernandezyopla.PianoReminder.Game.Repositories.EntityToSyncRepository;
 import com.dfernandezyopla.PianoReminder.Game.Repositories.HistoryQuestionRepository;
 import com.dfernandezyopla.PianoReminder.Game.Repositories.NoteRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
+import java.time.Instant;
 import java.util.*;
+import java.util.stream.Collectors;
 
 @Service
 public class QuestionServiceImpl implements QuestionService {
@@ -21,14 +28,17 @@ public class QuestionServiceImpl implements QuestionService {
     private final HistoryQuestionRepository historyQuestionRepository;
     private final NoteRepository noteRepository;
     private final ChordRepository chordRepository;
+    private final EntityToSyncRepository entityToSyncRepository;
 
     @Autowired
     public QuestionServiceImpl(HistoryQuestionRepository historyQuestionRepository,
                                NoteRepository noteRepository,
-                               ChordRepository chordRepository) {
+                               ChordRepository chordRepository,
+                               EntityToSyncRepository entityToSyncRepository) {
         this.historyQuestionRepository = historyQuestionRepository;
         this.noteRepository = noteRepository;
         this.chordRepository = chordRepository;
+        this.entityToSyncRepository = entityToSyncRepository;
     }
 
     @Override
@@ -37,18 +47,34 @@ public class QuestionServiceImpl implements QuestionService {
     }
 
     @Override
-    public List<Note> getNotes() {
-        return noteRepository.findAll();
-    }
-
-    @Override
-    public List<Chord> getChords() {
-        return chordRepository.findAll();
-    }
-
-    @Override
     public HistoryQuestion getHistoryQuestionById(Long id) {
         return historyQuestionRepository.findById(id).orElseThrow(() -> new HistoryQuestionNotFoundException("History question not found"));
+    }
+
+    @Override
+    public HistoryQuestion createHistoryQuestion(HistoryQuestion historyQuestion) {
+        HistoryQuestion createdQuestion = historyQuestionRepository.save(historyQuestion);
+        createEntityToSyncForCreation(createdQuestion.getId(), EntityType.HISTORY, SyncType.CREATED);
+        return createdQuestion;
+    }
+
+    @Override
+    public void deleteHistoryQuestion(Long id) {
+        if (!historyQuestionRepository.existsById(id)) {
+            throw new HistoryQuestionNotFoundException("Couldn't delete the question because Id doesn't exist");
+        }
+
+        try {
+            historyQuestionRepository.deleteById(id);
+            createEntityToSyncForCreation(id, EntityType.HISTORY, SyncType.DELETED);
+        } catch (Exception e) {
+            throw new RuntimeException("An unexpected error occurred while deleting history question with ID " + id, e);
+        }
+    }
+
+    @Override
+    public List<Note> getNotes() {
+        return noteRepository.findAll();
     }
 
     @Override
@@ -57,18 +83,36 @@ public class QuestionServiceImpl implements QuestionService {
     }
 
     @Override
+    public Note createNote(Note note) {
+        Note createdNote = noteRepository.save(note);
+        createEntityToSyncForCreation(createdNote.getId(), EntityType.NOTE, SyncType.CREATED);
+        return createdNote;
+    }
+
+    @Override
+    public void deleteNote(Long id) {
+        if (!noteRepository.existsById(id)) {
+            throw new HistoryQuestionNotFoundException("Couldn't delete the note because Id doesn't exist");
+        }
+
+        try {
+            noteRepository.deleteById(id);
+            createEntityToSyncForCreation(id, EntityType.NOTE, SyncType.DELETED);
+        } catch (DataIntegrityViolationException e) {
+            throw new EntityConstrainedException("This note belongs to another entity: " + e);
+        } catch (Exception e) {
+            throw new RuntimeException("An unexpected error occurred while deleting note with ID " + id, e);
+        }
+    }
+
+    @Override
+    public List<Chord> getChords() {
+        return chordRepository.findAll();
+    }
+
+    @Override
     public Chord getChordById(Long id) {
         return chordRepository.findById(id).orElseThrow(() -> new ChordNotFoundException("Chord not found"));
-    }
-
-    @Override
-    public HistoryQuestion createHistoryQuestion(HistoryQuestion historyQuestion) {
-        return historyQuestionRepository.save(historyQuestion);
-    }
-
-    @Override
-    public Note createNote(Note note) {
-        return noteRepository.save(note);
     }
 
     @Override
@@ -85,21 +129,37 @@ public class QuestionServiceImpl implements QuestionService {
         chord.setClef(chordDTO.getClef());
         chord.setTitle(chordDTO.getTitle());
         chord.setNotes(notes);
-        return chordRepository.save(chord);
-    }
 
-    @Override
-    public void deleteNote(Long id) {
-        noteRepository.deleteById(id);
+        Chord createdChord = chordRepository.save(chord);
+        createEntityToSyncForCreation(createdChord.getId(), EntityType.CHORD, SyncType.CREATED);
+        return createdChord;
     }
 
     @Override
     public void deleteChord(Long id) {
-        chordRepository.deleteById(id);
+        if (!chordRepository.existsById(id)) {
+            throw new ChordNotFoundException("Couldn't delete the chord because Id doesn't exist");
+        }
+
+        try {
+            chordRepository.deleteById(id);
+            createEntityToSyncForCreation(id, EntityType.CHORD, SyncType.DELETED);
+        } catch (Exception e) {
+            throw new RuntimeException("An unexpected error occurred while deleting chord with ID " + id, e);
+        }
     }
 
     @Override
-    public void deleteHistoryQuestion(Long id) {
-        historyQuestionRepository.deleteById(id);
+    public List<EntityToSync> getEntitiesToSync() {
+        return entityToSyncRepository.findAll();
+    }
+
+    private void createEntityToSyncForCreation(Long entityId, EntityType entityType, SyncType syncType) {
+        EntityToSync newEntity = new EntityToSync();
+        newEntity.setEntityId(entityId);
+        newEntity.setEntityType(entityType);
+        newEntity.setModifiedAt(Instant.now());
+        newEntity.setSyncType(syncType);
+        entityToSyncRepository.save(newEntity);
     }
 }
